게임 데이터 및 UI 작업 중

+++ Assets/3DRPG/Scripts/Character/Player/PlayerController.cs
@@ -8,7 +8,7 @@ public class PlayerController : MonoBehaviour
     [SerializeField] CharacterController controller;
     [SerializeField] GameObject playerRotation;
     [SerializeField] Transform camera_Point;
-    [SerializeField] float playerSpeed;
+    [SerializeField] int playerSpeed;
     [SerializeField] float acceleration;
     [SerializeField] float deceleration;
     [SerializeField] float turnSpeed;
